'Fix: 修复 "example" -> "examples" && "test" -> "tests"'
