--- example/example_radar.py
@@ -1,27 +1,0 @@
-from cutecharts.charts import Radar
-from cutecharts.components import Page
-from cutecharts.faker import Faker
-
-
-def radar_base() -> Radar:
-    chart = Radar("Radar-基本示例")
-    chart.set_options(labels=Faker.choose())
-    chart.add_series("series-A", Faker.values())
-    chart.add_series("series-B", Faker.values())
-    return chart
-
-
-radar_base().render()
-
-
-def radar_legend_colors():
-    chart = Radar("Radar-颜色调整")
-    chart.set_options(labels=Faker.choose(), colors=Faker.colors, legend_pos="upRight")
-    chart.add_series("series-A", Faker.values())
-    chart.add_series("series-B", Faker.values())
-    return chart
-
-
-page = Page()
-page.add(radar_base(), radar_legend_colors())
-page.render()
