+++ examples/example_bar.py
