+++ test.bat
@@ -1,2 +1,2 @@
-cd test
-nosetests --with-coverage --cover-package cutecharts --cover-package test && cd .. && flake8 --exclude build --max-line-length 89 --ignore=F401
+cd tests
+nosetests --with-coverage --cover-package cutecharts --cover-package tests && cd .. && flake8 --exclude build --max-line-length 89 --ignore=F401
