+++ tests/test_scatter.py
